avatar components ok > avatar update actions
